파이어베이스 DB 읽기, 날짜 수정 구현, 화면전환 수정
