+++ app/src/main/java/com/lion/FinalProject_CarryOn_Anywhere/ui/screen/trip/SelectTripDateScreen.kt
@@ -1,5 +1,6 @@
 package com.lion.FinalProject_CarryOn_Anywhere.ui.screen.trip
 
+import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
@@ -16,35 +17,70 @@ import com.lion.FinalProject_CarryOn_Anywhere.component.LikeLionFilledButton
 import com.lion.FinalProject_CarryOn_Anywhere.component.LikeLionTopAppBar
 import com.lion.FinalProject_CarryOn_Anywhere.ui.theme.GrayColor
 import com.lion.FinalProject_CarryOn_Anywhere.ui.theme.SubColor
+import com.lion.FinalProject_CarryOn_Anywhere.ui.viewmodel.trip.AddTripInfoViewModel
 import com.lion.FinalProject_CarryOn_Anywhere.ui.viewmodel.trip.TripInfoViewModel
 import java.util.Calendar
+import java.util.Locale
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun SelectTripDateScreen(
-    tripInfoViewModel: TripInfoViewModel = hiltViewModel(),
+    addTripInfoViewModel: AddTripInfoViewModel,
     tripDocumentId: String?
 ) {
-    if (!tripDocumentId.isNullOrEmpty()) {
-        tripInfoViewModel.gettingTripData(tripDocumentId)
+    LaunchedEffect(tripDocumentId) {
+        if (!tripDocumentId.isNullOrEmpty()) {
+            try {
+                Log.d("SelectTripDateScreen", "📡 서버에서 데이터 가져오기: tripDocumentId = $tripDocumentId")
+                addTripInfoViewModel.gettingTripData(tripDocumentId)
+            } catch (e: Exception) {
+                Log.e("SelectTripDateScreen", "🚨 데이터 가져오기 실패: ${e.message}")
+            }
+        }
+    }
+
+    // ✅ 서버에서 데이터를 가져온 후 UI를 최신화
+    LaunchedEffect(addTripInfoViewModel.serverStartDate.value, addTripInfoViewModel.serverEndDate.value) {
+        if (addTripInfoViewModel.serverStartDate.value != null) {
+            Log.d("SelectTripDateScreen", "✅ 서버 데이터 로드 완료: ${addTripInfoViewModel.serverStartDate.value} ~ ${addTripInfoViewModel.serverEndDate.value}")
+
+            // 서버에서 받아온 날짜를 화면의 선택 날짜로 설정
+            addTripInfoViewModel.startDate.value = addTripInfoViewModel.serverStartDate.value
+            addTripInfoViewModel.endDate.value = addTripInfoViewModel.serverEndDate.value
+            addTripInfoViewModel.updateFormattedDates()
+        }
     }
 
     val calendar = Calendar.getInstance()
-
-    // 현재 월의 시작과 끝 설정
     val startOfMonthMillis = calendar.apply { set(Calendar.DAY_OF_MONTH, 1) }.timeInMillis
 
-    // `DateRangePicker`에서 현재 월까지만 선택 가능하도록 설정
-    val dateRangePickerState = rememberDateRangePickerState(
-        initialDisplayedMonthMillis = startOfMonthMillis, // 현재 월부터 시작
-        initialSelectedStartDateMillis = System.currentTimeMillis(), // 출발 날짜 초기화
-    )
+    // ✅ 서버에서 받아온 값이 변경될 때 `DateRangePickerState`를 강제로 다시 생성
+    var dateRangePickerState by remember {
+        mutableStateOf(
+            DateRangePickerState(
+                initialDisplayedMonthMillis = startOfMonthMillis,
+                initialSelectedStartDateMillis = addTripInfoViewModel.serverStartDate.value ?: System.currentTimeMillis(),
+                initialSelectedEndDateMillis = addTripInfoViewModel.serverEndDate.value,
+                locale = Locale.KOREA
+            )
+        )
+    }
 
-    // 선택한 날짜가 변경될 때 `ViewModel` 업데이트
+    // ✅ 서버 값이 변경될 때 `DateRangePickerState`를 재생성
+    LaunchedEffect(addTripInfoViewModel.serverStartDate.value, addTripInfoViewModel.serverEndDate.value) {
+        dateRangePickerState = DateRangePickerState(
+            initialDisplayedMonthMillis = startOfMonthMillis,
+            initialSelectedStartDateMillis = addTripInfoViewModel.serverStartDate.value ?: System.currentTimeMillis(),
+            initialSelectedEndDateMillis = addTripInfoViewModel.serverEndDate.value,
+            locale = Locale.KOREA
+        )
+    }
+
+    // ✅ 사용자가 날짜를 변경할 때 UI 상태 업데이트
     LaunchedEffect(dateRangePickerState.selectedStartDateMillis, dateRangePickerState.selectedEndDateMillis) {
-        tripInfoViewModel.startDate.value = dateRangePickerState.selectedStartDateMillis
-        tripInfoViewModel.endDate.value = dateRangePickerState.selectedEndDateMillis
-        tripInfoViewModel.updateFormattedDates()
+        addTripInfoViewModel.startDate.value = dateRangePickerState.selectedStartDateMillis
+        addTripInfoViewModel.endDate.value = dateRangePickerState.selectedEndDateMillis
+        addTripInfoViewModel.updateFormattedDates()
     }
 
     Scaffold(
@@ -53,7 +89,7 @@ fun SelectTripDateScreen(
                 title = "여행 일정 선택",
                 navigationIconImage = ImageVector.vectorResource(R.drawable.arrow_back_24px),
                 navigationIconOnClick = {
-                    tripInfoViewModel.tripDateNavigationOnClick(tripDocumentId ?: "")
+                    addTripInfoViewModel.tripDateNavigationOnClick(tripDocumentId ?: "")
                 }
             )
         }
@@ -72,15 +108,36 @@ fun SelectTripDateScreen(
                 modifier = Modifier.padding(top = 50.dp)
             )
 
-            // `DateRangePicker`에서 현재 월까지만 선택 가능하도록 설정
             DateRangePicker(
                 title = { Text("") },
                 headline = {
                     Text(
-                        text = if (tripInfoViewModel.endDate.value == null || tripInfoViewModel.formattedStartDate == tripInfoViewModel.formattedEndDate) {
-                            tripInfoViewModel.formattedStartDate.value
+                        text = if (tripDocumentId.isNullOrEmpty()) {
+                            if (addTripInfoViewModel.endDate.value == null ||
+                                addTripInfoViewModel.formattedStartDate == addTripInfoViewModel.formattedEndDate) {
+                                addTripInfoViewModel.formattedStartDate.value
+                            } else {
+                                "${addTripInfoViewModel.formattedStartDate.value} ~ ${addTripInfoViewModel.formattedEndDate.value}"
+                            }
                         } else {
-                            "${tripInfoViewModel.formattedStartDate.value} ~ ${tripInfoViewModel.formattedEndDate.value}"
+                            val startChanged = addTripInfoViewModel.formattedStartDate.value != addTripInfoViewModel.formattedServerStartDate.value
+                            val endChanged = addTripInfoViewModel.formattedEndDate.value != addTripInfoViewModel.formattedServerEndDate.value
+
+                            if (startChanged || endChanged) {
+                                if (addTripInfoViewModel.endDate.value == null ||
+                                    addTripInfoViewModel.formattedStartDate == addTripInfoViewModel.formattedEndDate) {
+                                    addTripInfoViewModel.formattedStartDate.value
+                                } else {
+                                    "${addTripInfoViewModel.formattedStartDate.value} ~ ${addTripInfoViewModel.formattedEndDate.value}"
+                                }
+                            } else {
+                                if (addTripInfoViewModel.serverEndDate.value == null ||
+                                    addTripInfoViewModel.formattedServerStartDate == addTripInfoViewModel.formattedServerEndDate) {
+                                    addTripInfoViewModel.formattedServerStartDate.value
+                                } else {
+                                    "${addTripInfoViewModel.formattedServerStartDate.value} ~ ${addTripInfoViewModel.formattedServerEndDate.value}"
+                                }
+                            }
                         },
                         textAlign = TextAlign.Start,
                         modifier = Modifier.padding(bottom = 15.dp)
@@ -101,35 +158,26 @@ fun SelectTripDateScreen(
                 )
             )
 
-            if (tripDocumentId.isNullOrEmpty()) {
-                // 새 여행 등록
-                LikeLionFilledButton(
-                    text = if (tripInfoViewModel.endDate.value == null || tripInfoViewModel.formattedStartDate.value == tripInfoViewModel.formattedEndDate.value) {
-                        "${tripInfoViewModel.formattedStartDate.value} 등록완료"
+            LikeLionFilledButton(
+                text = if (addTripInfoViewModel.formattedEndDate.value.isNullOrEmpty()) {
+                    "${addTripInfoViewModel.formattedStartDate.value} ${
+                        if (tripDocumentId.isNullOrEmpty()) " 등록완료" else " 수정완료"
+                    }"
+                } else {
+                    "${addTripInfoViewModel.formattedStartDate.value} ~ ${addTripInfoViewModel.formattedEndDate.value} ${
+                        if (tripDocumentId.isNullOrEmpty()) " 등록완료" else " 수정완료"
+                    }"
+                },
+                onClick = {
+                    if (tripDocumentId.isNullOrEmpty()) {
+                        addTripInfoViewModel.completeDateOnClick()
                     } else {
-                        "${tripInfoViewModel.formattedStartDate.value} ~ ${tripInfoViewModel.formattedEndDate.value} 등록완료"
-                    },
-                    onClick = {
-                        tripInfoViewModel.completeDateOnClick()
-                    },
-                    modifier = Modifier.fillMaxWidth().padding(bottom = 5.dp),
-                    cornerRadius = 5
-                )
-            } else {
-                // 기존 여행 일정 수정
-                LikeLionFilledButton(
-                    text = if (tripInfoViewModel.endDate.value == null || tripInfoViewModel.formattedStartDate.value == tripInfoViewModel.formattedEndDate.value) {
-                        "${tripInfoViewModel.formattedStartDate.value} 수정완료"
-                    } else {
-                        "${tripInfoViewModel.formattedStartDate.value} ~ ${tripInfoViewModel.formattedEndDate.value} 등록완료"
-                    },
-                    onClick = {
-                        tripInfoViewModel.updateDateOnClick(tripDocumentId)
-                    },
-                    modifier = Modifier.fillMaxWidth().padding(bottom = 5.dp),
-                    cornerRadius = 5
-                )
-            }
+                        addTripInfoViewModel.updateDateOnClick(tripDocumentId)
+                    }
+                },
+                modifier = Modifier.fillMaxWidth().padding(bottom = 5.dp),
+                cornerRadius = 5
+            )
         }
     }
 }
